backend/requirements.txt
# ==============================================================================
# ResXiv V2 Backend Dependencies
# ==============================================================================

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.31.1
python-multipart==0.0.12

# ASGI Server and HTTP
starlette==0.41.0
httpx==0.27.2

# Database - PostgreSQL
asyncpg==0.29.0
psycopg2-binary==2.9.10
sqlalchemy[asyncio]==2.0.35
alembic==1.13.3

# Database - MongoDB
motor==3.6.0
pymongo==4.10.1

# Database - Redis
redis==5.1.1
aioredis==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
cryptography==43.0.1

# Environment & Configuration
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0

# AI & ML
openai==1.54.4
langchain==0.3.7
langgraph==0.2.34
langsmith==0.1.139

# Research APIs & Data Processing
requests==2.32.3
beautifulsoup4==4.12.3
feedparser==1.6.3
arxiv==2.1.3

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.11.4
pymupdf==1.24.13

# File Processing & Git
GitPython==3.1.43
pathlib2==2.3.7

# Email
fastapi-mail==1.4.1
aiosmtplib==3.0.2

# Rate Limiting & Middleware
slowapi==0.1.9

# Data Processing & Utilities
pandas==2.2.3
numpy==1.26.4
scipy==1.14.1

# JSON & Data Serialization
orjson==3.10.7

# Logging & Monitoring
structlog==24.4.0

# Testing (Development)
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2  # for testing API endpoints

# Code Quality (Development)
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.13.0

# Documentation
mkdocs==1.6.1
mkdocs-material==9.5.41

# LaTeX & Document Processing
python-magic==0.4.27

# Image Processing
Pillow==10.4.0

# Date & Time
python-dateutil==2.9.0

# Async utilities
aiofiles==24.1.0

# Data validation
validators==0.34.0

# UUID utilities
uuid==1.30

# Type hints backport (Python < 3.10)
typing-extensions==4.12.2 